etl/requirements.txt
geopandas
sqlalchemy
psycopg2-binary
geoalchemy2
shapely
fiona
pyproj
openai
python-dotenv